+++ modules/scheduler/scheduler/pom.xml
@@ -139,10 +139,11 @@
                         <Blueprint-Enabled>true</Blueprint-Enabled>
                         <Export-Package>
                             org.motechproject.scheduler.builder;version=${project.version},
-                            org.motechproject.scheduler.service.*;version=${project.version},
                             org.motechproject.scheduler.contract;version=${project.version},
+                            org.motechproject.scheduler.exception;version=${project.version},
                             org.motechproject.scheduler.factory;version=${project.version},
-                            org.motechproject.scheduler.exception;version=${project.version}
+                            org.motechproject.scheduler.service;version=${project.version},
+                            org.motechproject.scheduler.tasks;version=${project.version}
                         </Export-Package>
                         <Import-Package>
                             org.apache.commons.dbcp,
@@ -257,7 +258,7 @@
                     <transformers>
                         <transformer implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
                             <manifestEntries>
-                                <Main-Class>org.motechproject.scheduler.service.impl.MotechScheduler</Main-Class>
+                                <Main-Class>org.motechproject.scheduler.service.MotechScheduler</Main-Class>
                             </manifestEntries>
                         </transformer>
                         <transformer implementation="org.apache.maven.plugins.shade.resource.IncludeResourceTransformer">
