+++ modules/scheduler/scheduler/src/main/java/org/motechproject/scheduler/constants/SchedulerConstants.java
@@ -1,7 +1,7 @@
 package org.motechproject.scheduler.constants;
 
 /**
- * Contains constants used for securing parts of the scheduler module.
+ * Contains constants used for the scheduler module.
  */
 public final class SchedulerConstants {
 
@@ -26,6 +26,9 @@ public final class SchedulerConstants {
     public static final String REPEAT_INTERVAL_IN_SECONDS = "repeatIntervalInMilliseconds";
     public static final String IGNORE_PAST_FIRES_AT_START = "ignorePastFiresAtStart";
     public static final String USE_ORIGINAL_FIRE_TIME_AFTER_MISFIRE = "useOriginalFireTimeAfterMisfire";
+
+    /* Scheduler tasks */
+    public static final String PARSER_NAME = "SchedulerJobs";
 
     private SchedulerConstants() {
