+++ modules/scheduler/scheduler/src/main/java/org/motechproject/scheduler/service/MotechScheduledJob.java
@@ -1,11 +1,9 @@
-package org.motechproject.scheduler.service.impl;
+package org.motechproject.scheduler.service;
 
 import org.motechproject.commons.api.TasksEventParser;
 import org.motechproject.event.MotechEvent;
 import org.motechproject.event.listener.EventRelay;
 import org.motechproject.scheduler.constants.SchedulerConstants;
-import org.motechproject.scheduler.service.MotechSchedulerService;
-import org.motechproject.scheduler.tasks.SchedulerEventParser;
 import org.quartz.Job;
 import org.quartz.JobDataMap;
 import org.quartz.JobDetail;
@@ -49,7 +47,7 @@ public class MotechScheduledJob implements Job {
             Map<String, Object> params = jobDataMap.getWrappedMap();
             params.remove(SchedulerConstants.EVENT_TYPE_KEY_NAME);
             params.put(MotechSchedulerService.JOB_ID_KEY, jobId);
-            params.put(TasksEventParser.CUSTOM_PARSER_EVENT_KEY, SchedulerEventParser.PARSER_NAME);
+            params.put(TasksEventParser.CUSTOM_PARSER_EVENT_KEY, SchedulerConstants.PARSER_NAME);
 
             MotechEvent motechEvent = new MotechEvent(eventType, params);
