+++ modules/scheduler/scheduler/src/main/java/org/motechproject/scheduler/tasks/DynamicChannelProviderRegisterer.java
@@ -1,6 +1,5 @@
 package org.motechproject.scheduler.tasks;
 
-import org.motechproject.scheduler.service.SchedulerChannelProvider;
 import org.motechproject.tasks.domain.DynamicChannelProvider;
 import org.osgi.framework.Bundle;
 import org.osgi.framework.BundleContext;
