+++ modules/scheduler/scheduler/src/test/java/org/motechproject/scheduler/tasks/DynamicChannelProviderRegistererTest.java
@@ -4,7 +4,6 @@ import org.junit.Before;
 import org.junit.Test;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
-import org.motechproject.scheduler.service.SchedulerChannelProvider;
 import org.motechproject.tasks.domain.DynamicChannelProvider;
 import org.osgi.framework.Bundle;
 import org.osgi.framework.BundleContext;
